improve types
