+++ src/components/Footer/Footer.tsx
@@ -1,13 +1,8 @@
 import BottomNavigation from "@material-ui/core/BottomNavigation";
 import React from "react";
 
-//TODO fix footer
-const Footer =()=>{
-  return (
-    <BottomNavigation>
-
-    </BottomNavigation>
-  )
-}
+const Footer = () => {
+  return <BottomNavigation></BottomNavigation>;
+};
 
 export default Footer;
