+++ src/components/SettingsPage/SettingsPage.tsx
@@ -9,7 +9,7 @@ import Field from "../Field/Field";
 import { IFoundation } from "../../interfaces/Foundation";
 import InformationTable from "../InformationTable/InformationTable";
 import GamePlayHandler from "../../services/GamePlayHandler";
-import { ICell } from "../../interfaces/Matrix";
+import { ICell } from "../../interfaces/Cell";
 import {
   Dialog,
   DialogTitle,
