+++ src/interfaces/Field.ts
@@ -1,0 +1,4 @@
+export interface IFieldProps {
+  foundation: number;
+  matrix: any;
+}
