+++ src/interfaces/InformationTable.ts
@@ -1,0 +1,6 @@
+export interface IInformationTableProps {
+  isFirstPlayerTurn: boolean;
+  firstPlayerSequencesLength: number;
+  secondPlayerSequencesLength: number;
+  isGameFinished: boolean;
+}
